+++ gmf_examples/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Features
 
-- Add play http music example
+- Added play http music example
 - Added `pipeline_http_download_to_sdcard` example for HTTP file downloads to SD card
+- Added `pipeline_record_http` example for uploading the recorded audio to HTTP server
 
 ## v0.7.2~1